main updates

+++ ngs-pipeline.sh
@@ -11,7 +11,8 @@ CFG="${SCRIPTS}/ngs-pipeline-${REF}.config"
 [[ -z "$SAMPLE" ]] && { echo "ERROR: No SAMPLE provided for this run"; exit 1; }
 [[ -z "$REF" ]] && { echo "ERROR: No REFERENCE provided for this run"; exit 1; }
 
-date
+date +"%d/%m/%Y %H:%M"
+echo $SAMPLE $REF;
 
 PCR_MODEL='CONSERVATIVE';
 
@@ -28,8 +29,9 @@ fi
 mkdir -p $SAMPLE; cd $SAMPLE
 cp $CFG ${SAMPLE}-${REF}.config; source ${SAMPLE}-${REF}.config
 
+
 # Copy files from RCS... will only copy if files no present or rcs version is newer
-rsync --progress -av ${WGS}/${SAMPLE}/*.fq.gz ./
+rsync --progress -av --no-links ${WGS}/${SAMPLE}/*.fq.gz ./
 
 # Count how many fastq files we have
 COUNT=`ls *.fq.gz | wc -l`
@@ -41,17 +43,34 @@ if [ $LANES -le 1 ]; then
   exit 1;
 fi
 
+ls *r_1.fq.gz > fastq.files
+if [ ! -s fastq.files ];then
+  echo ; echo "ERROR - The FASTQ filenames are not in the required format. Please rename them to end in r_1/r_2.fq.gz"; echo
+  exit 1;
+fi
+
+
 mkdir -p $GENOME; 
 mv ${SAMPLE}-${REF}.config $GENOME/${SAMPLE}.config
 cd $GENOME
 
+scp kcgc:/home/kcgc/tmp/sampleLookups.txt /home/es904/tmp/sampleLookups.txt
+grep -i ${SAMPLE}  /home/es904/tmp/sampleLookups.txt | awk '{ print "sample\t" $1 "\nwgsID\t" $3 "\nBioSample\t" $4 }' > ${SAMPLE}-${REF}.stats
+echo -e "refGenome\t${REF}" >> ${SAMPLE}-${REF}.stats
+echo -e "dateSubmitted\t"`date +"%Y-%m-%d %H:%M:%S"` >> ${SAMPLE}-${REF}.stats
+
 mkdir metrics logs intervals base_recal gvcf
 
 # count raw reads
-sbatch -A ${ACCOUNT} -J ${SAMPLE}.countReads ${SCRIPTS}/slurm/countRawReads.sh ${SAMPLE}
+sbatch -A ${ACCOUNT} -J ${SAMPLE}.countReads ${SCRIPTS}/slurm/countRawReads.sh ${SAMPLE} ${REF}
 
 module load ${GATK}
 gatk SplitIntervals -R ${FASTA}/${GENOME}.fasta -L ${INTERVAL_LIST} --scatter-count ${INTERVALS} -O intervals --subdivision-mode BALANCING_WITHOUT_INTERVAL_SUBDIVISION
+
+#BALANCING_WITHOUT_INTERVAL_SUBDIVISION means that --scatter-count is the upper limit for count of intervals
+#Need to reassign INTERVALS to the number actually generated
+INTERVALS=`ls intervals | wc -l`
+sed -i s/INTERVALS=50/INTERVALS=${INTERVALS}/ ${SAMPLE}.config
 
 # fastq2sam
 jid1=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.fastq2sam --array=1-${LANES} ${SCRIPTS}/slurm/make-bam/fastq2sam.sh ${SAMPLE})
@@ -91,7 +110,7 @@ jid9=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.gatherBQSR --dependency=afterok:${jid8#
 jid10=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.applyBQSR --dependency=afterok:${jid9##* } --array=0-$(($INTERVALS-1)) ${SCRIPTS}/slurm/make-bam/applyBSQR.sh ${SAMPLE})
 
 # Combine multiple recalibrated BAM files from scattered ApplyRecalibration runs - removes base_recal dir, ${SAMPLE}.sorted.bam, ${SAMPLE}.sorted.bai, ${SAMPLE}.sorted.bam.md5 & ${SAMPLE}.bsqr.out
-jid11=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.BAM --dependency=afterok:${jid10##* } ${SCRIPTS}/slurm/make-bam/bqsrMerge.sh ${SAMPLE} ${INTERVALS} ${REF})
+jid11=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.BAM --dependency=afterok:${jid10##* } ${SCRIPTS}/slurm/make-bam/bqsrMerge.sh ${SAMPLE} ${REF})
 
 
 ###########################################################
@@ -119,9 +138,26 @@ jid13=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.HC --dependency=afterok:${jid11##* } -
 #jid14=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.GVCF --dependency=afterok:${jid13##* } ${SCRIPTS}/slurm/make-gvcf/combineGvcf.sh ${SAMPLE} ${INTERVALS} ${REF})
 jid14=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.GVCF --dependency=afterok:${jid13##* } ${SCRIPTS}/slurm/make-gvcf/mergeGvcfs.sh ${SAMPLE})
 
+
+###########################################################
+#                     VCF Files/Stats                     #
+###########################################################
+
+# Convert gVCF file to VCF
+jid15=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.VCF --dependency=afterok:${jid14##* } ${SCRIPTS}/slurm/gvcf2vcf.sh ${SAMPLE} ${REF})
+
+# Filter VCF and split into snps/indels
+jid16=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.filterVCF --dependency=afterok:${jid15##* } ${SCRIPTS}/slurm/filterVcf.sh ${SAMPLE} ${REF})
+
+# Output VCF stats
+tmp=$(sbatch -A ${ACCOUNT} -J ${SAMPLE}.VCFstats --dependency=afterok:${jid16##* } ${SCRIPTS}/slurm/vcfStats.sh ${SAMPLE} ${REF})
+
+
+
 ###########################################################
 #                         OUTPUT                          #
 ###########################################################
 echo "BAM file = job-${jid11##* }"
 echo "gVCF file = job-${jid14##* }"
+echo "VCF file = job-${jid16##* }"
 echo "bamMetrics-${jid12##* }"
